LayoutTests/imported/w3c/web-platform-tests/pointerevents/extension/pointerevent_touch-action-verification-expected.txt

Pointer Events touch-action attribute support

Test Description: Test will automatically check parsing behaviour of various touch-action combinations.

touch-action: basic verification

The following pointer types were detected: .


FAIL default assert_equals: expected (undefined) undefined but got (string) "auto"
FAIL stylesheet-none assert_equals: expected (undefined) undefined but got (string) "none"
FAIL explicit-auto assert_equals: expected (undefined) undefined but got (string) "auto"
FAIL explicit-pan-x assert_equals: expected (undefined) undefined but got (string) "pan-x"
FAIL explicit-pan-left assert_equals: expected (undefined) undefined but got (string) "pan-left"
FAIL explicit-pan-right assert_equals: expected (undefined) undefined but got (string) "pan-right"
FAIL explicit-pan-y assert_equals: expected (undefined) undefined but got (string) "pan-y"
FAIL explicit-pan-up assert_equals: expected (undefined) undefined but got (string) "pan-up"
FAIL explicit-pan-down assert_equals: expected (undefined) undefined but got (string) "pan-down"
FAIL explicit-pinch-zoom assert_equals: expected (undefined) undefined but got (string) "pinch-zoom"
FAIL explicit-pan-x-pan-y assert_equals: expected (undefined) undefined but got (string) "pan-x pan-y"
FAIL explicit-pan-y-pan-x assert_equals: expected (undefined) undefined but got (string) "pan-x pan-y"
FAIL explicit-pan-left-pan-up assert_equals: expected (undefined) undefined but got (string) "pan-left pan-up"
FAIL explicit-pan-left-pan-down assert_equals: expected (undefined) undefined but got (string) "pan-left pan-down"
FAIL explicit-pan-right-pan-up assert_equals: expected (undefined) undefined but got (string) "pan-right pan-up"
FAIL explicit-pan-right-pan-down assert_equals: expected (undefined) undefined but got (string) "pan-right pan-down"
FAIL explicit-pan-up-pan-left assert_equals: expected (undefined) undefined but got (string) "pan-left pan-up"
FAIL explicit-pan-up-pan-right assert_equals: expected (undefined) undefined but got (string) "pan-right pan-up"
FAIL explicit-pan-down-pan-left assert_equals: expected (undefined) undefined but got (string) "pan-left pan-down"
FAIL explicit-pan-down-pan-right assert_equals: expected (undefined) undefined but got (string) "pan-right pan-down"
FAIL explicit-pinch-zoom-pan-x-pan-up assert_equals: expected (undefined) undefined but got (string) "pan-x pan-up pinch-zoom"
FAIL explicit-pinch-zoom-pan-x-pan-y assert_equals: expected (undefined) undefined but got (string) "manipulation"
FAIL explicit-manipulation assert_equals: expected (undefined) undefined but got (string) "manipulation"
FAIL explicit-none assert_equals: expected (undefined) undefined but got (string) "none"
FAIL explicit-invalid-1 assert_equals: expected (undefined) undefined but got (string) "auto"
FAIL explicit-invalid-2 assert_equals: expected (undefined) undefined but got (string) "none"
FAIL explicit-invalid-3 assert_equals: expected (undefined) undefined but got (string) "auto"
FAIL explicit-invalid-4 assert_equals: expected (undefined) undefined but got (string) "auto"
FAIL explicit-invalid-5 assert_equals: expected (undefined) undefined but got (string) "auto"
FAIL explicit-invalid-6 assert_equals: expected (undefined) undefined but got (string) "auto"
FAIL explicit-invalid-7 assert_equals: expected (undefined) undefined but got (string) "auto"
FAIL explicit-invalid-8 assert_equals: expected (undefined) undefined but got (string) "auto"
FAIL explicit-invalid-9 assert_equals: expected (undefined) undefined but got (string) "auto"
FAIL explicit-invalid-10 assert_equals: expected (undefined) undefined but got (string) "auto"
FAIL explicit-invalid-11 assert_equals: expected (undefined) undefined but got (string) "auto"
FAIL explicit-invalid-12 assert_equals: expected (undefined) undefined but got (string) "auto"
FAIL explicit-invalid-13 assert_equals: expected (undefined) undefined but got (string) "auto"
FAIL explicit-invalid-14 assert_equals: expected (undefined) undefined but got (string) "auto"
FAIL not-inherited assert_equals: expected (undefined) undefined but got (string) "auto"
FAIL inherit assert_equals: expected (undefined) undefined but got (string) "none"
FAIL initial assert_equals: expected (undefined) undefined but got (string) "auto"
